requirements.txt
Django>=4.2,<6
Pillow>=10.0.0
